+++ app/src/androidTest/java/com/twilio/video/app/espresso/DrawableMatchers.kt
@@ -2,6 +2,9 @@ package com.twilio.video.app.espresso
 
 import android.content.Context
 import android.content.res.Resources
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.Drawable
 import android.view.View
 import android.widget.ImageView
@@ -25,11 +28,30 @@ class DrawableMatcher(
 
         val resources: Resources = target.context.resources
         val expectedDrawable: Drawable? = resources.getDrawable(expectedId, targetContext.theme)
-        return expectedDrawable?.constantState?.let { it == drawable.constantState } ?: false
+        return if (expectedDrawable != null) {
+            makeBitmap(drawable)?.sameAs(makeBitmap(expectedDrawable)) ?: false
+        } else {
+            false
+        }
     }
 
     override fun describeTo(description: Description) {
         description.appendText("with drawable from resource id: $expectedId")
         targetContext.resources.getResourceEntryName(expectedId)?.let { description.appendText("[$it]") }
+    }
+
+    private fun makeBitmap(drawable: Drawable): Bitmap? {
+        if (drawable is BitmapDrawable) {
+            return drawable.bitmap
+        }
+        val bitmap = Bitmap.createBitmap(
+            drawable.intrinsicWidth,
+            drawable.intrinsicHeight,
+            Bitmap.Config.ARGB_8888
+        )
+        val canvas = Canvas(bitmap)
+        drawable.setBounds(0, 0, canvas.width, canvas.height)
+        drawable.draw(canvas)
+        return bitmap
     }
 }