+++ app/src/main/java/com/twilio/video/app/auth/GoogleAuthProvider.kt
@@ -65,6 +65,7 @@ class GoogleAuthProvider @JvmOverloads internal constructor(
                             }
                         }
                     }
+                    else -> { Timber.e("Should never reach here") }
                 }
             },
                     {