+++ app/src/main/java/com/twilio/video/app/data/api/dto/AuthServiceResponseDTO.kt
@@ -1,7 +1,7 @@
 package com.twilio.video.app.data.api
 
 import com.google.gson.annotations.SerializedName
-import com.twilio.video.app.data.api.model.Topology
+import com.twilio.video.app.data.api.dto.Topology
 
 data class AuthServiceResponseDTO(
     val token: String? = null,