+++ app/src/main/java/com/twilio/video/app/data/api/dto/Topology.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.twilio.video.app.data.api.model
+package com.twilio.video.app.data.api.dto
 
 import com.google.gson.annotations.SerializedName